+++ out/extension.js
@@ -41,26 +41,36 @@ const diagnosticProvider_1 = require("./diagnosticProvider");
 let quoteConverter;
 let diagnosticProvider;
 let warningDiagnosticProvider;
+let commaDiagnosticProvider;
+let spaceDiagnosticProvider;
 function activate(context) {
     console.log('ZS language extension is now active!');
     quoteConverter = new quotationMarkConverter_1.ZSQuoteConverter();
     diagnosticProvider = new diagnosticProvider_1.ZSDiagnosticProvider(context);
     warningDiagnosticProvider = new diagnosticProvider_1.ZSDiagnosticProviderWarning();
+    commaDiagnosticProvider = new diagnosticProvider_1.ZSCommaDiagnosticProvider();
+    spaceDiagnosticProvider = new diagnosticProvider_1.ZSSpaceDiagnosticProvider();
     quoteConverter.activate(context);
     const diagnosticDisposable = vscode.workspace.onDidChangeTextDocument((event) => {
         diagnosticProvider.updateDiagnostics(event.document);
         warningDiagnosticProvider.updateDiagnostics(event.document);
+        commaDiagnosticProvider.updateDiagnostics(event.document);
+        spaceDiagnosticProvider.updateDiagnostics(event.document);
     });
     const openDisposable = vscode.workspace.onDidOpenTextDocument((document) => {
         if (document.languageId === 'zs') {
             diagnosticProvider.updateDiagnostics(document);
             warningDiagnosticProvider.updateDiagnostics(document);
+            commaDiagnosticProvider.updateDiagnostics(document);
+            spaceDiagnosticProvider.updateDiagnostics(document);
         }
     });
     vscode.workspace.textDocuments.forEach(document => {
         if (document.languageId === 'zs') {
             diagnosticProvider.updateDiagnostics(document);
             warningDiagnosticProvider.updateDiagnostics(document);
+            commaDiagnosticProvider.updateDiagnostics(document);
+            spaceDiagnosticProvider.updateDiagnostics(document);
         }
     });
     const commandDisposable = vscode.commands.registerCommand('zs.convertQuotes', () => {
@@ -76,6 +86,12 @@ function activate(context) {
             if (warningDiagnosticProvider) {
                 warningDiagnosticProvider.dispose();
             }
+            if (commaDiagnosticProvider) {
+                commaDiagnosticProvider.dispose();
+            }
+            if (spaceDiagnosticProvider) {
+                spaceDiagnosticProvider.dispose();
+            }
         } }, vscode.languages.registerCodeActionsProvider('zs', diagnosticProvider, {
         providedCodeActionKinds: [vscode.CodeActionKind.QuickFix]
     }));
@@ -89,6 +105,12 @@ function deactivate() {
     }
     if (warningDiagnosticProvider) {
         warningDiagnosticProvider.dispose();
+    }
+    if (commaDiagnosticProvider) {
+        commaDiagnosticProvider.dispose();
+    }
+    if (spaceDiagnosticProvider) {
+        spaceDiagnosticProvider.dispose();
     }
 }
 //# sourceMappingURL=extension.js.map