+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Change Log
-- Update `diagnosticProvider.ts` to skip symbols: `-/`, `*/-` and `/-*`, `<>`, `/||\`, `()`, `‘’`, `“”`
-- Up to version 0.1.9
+- Update `diagnosticProvider.ts`:
+    + Add "Commas Detector" to underline `,` with red underline for `<>`, `()`, `/||\`
+    + Add "Spaces Detector" to underline ` ` with red underline for `<>`, `()`, `/||\`
+- Up to version 0.1.10